+++ rmk/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add layout macro to! and df!
+
 ### Changed
 
 - Use `User11` to manually switch between USB mode and BLE mode